+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# TBD
+
++ Fix virtualbox int tests
++ Fix bug in "up" operation for virtualbox
++ Guess which network interface to use for bridge interface ("no-nat")
++ Add "purge" option to "mech global-status"
+
 # v0.8.7
 
 + make python3.7+ requirement